ImageView and ScrollView implemented
ImageView and ScrollView implemented somewhat correctly - the aspect ratio of the picture stays correct and the Clip View cell is exactly the size of the window picture is displayed in.
However, there is one issue - when zooming out the picture gets stuck on the bottom left corner of the window.

Still have to implement functionfeature which lets open any image (like open any image after the code was compiled).