Fixed the coordinates and the size of a newly opened window
Created two new functons in Document.swift - origin and winsizeXY.
Origin simply returns such coordinates of bottom left corner of the window that it is right in the middle of the screen.
WinsizeXY returns the size of the window. No larger than the screen itself but no smaller than 150x300